test(testkit): separate MCP OS contracts
